move keycloak auth logic to its own class
