+++ app/src/main/java/com/feedhenry/securenativeandroidtemplate/authenticate/AuthenticateResult.java
@@ -4,7 +4,19 @@ package com.feedhenry.securenativeandroidtemplate.authenticate;
  * Created by weili on 04/09/2017.
  */
 
-//TODO: implement me!
 public class AuthenticateResult {
 
+    private boolean result;
+
+    public AuthenticateResult(boolean result) {
+        this.result = result;
+    }
+
+    public boolean getResult() {
+        return result;
+    }
+
+    public void setResult(boolean result) {
+        this.result = result;
+    }
 }
